requirements.txt
amqp==5.3.1
anyio==4.11.0
APScheduler==3.11.0
asgiref==3.9.1
asttokens==3.0.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
decorator==5.2.1
Django==5.2.4
django-environ==0.12.0
django-extensions==4.1
django-leaflet==0.32.0
django-qrcode==0.3
django-widget-tweaks==1.5.0
executing==2.2.0
flatbuffers==25.2.10
h11==0.16.0
h3==4.3.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.1.7
numpy==2.3.2
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-telegram-bot==20.7
redis==6.4.0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
timezonefinder==7.0.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
